requirements.txt
aiodns==3.0.0; python_version >= "3.6" and python_full_version >= "3.8.0"
aiofiles==0.8.0; python_version >= "3.6" and python_version < "4.0"
aiohttp==3.8.1; python_version >= "3.6"
aiosignal==1.2.0; python_version >= "3.6" and python_full_version >= "3.8.0"
async-timeout==4.0.2; python_version >= "3.6" and python_full_version >= "3.8.0"
attrs==22.1.0; python_version >= "3.6" and python_full_version >= "3.8.0"
bcrypt==3.2.2; python_version >= "3.6"
bidict==0.22.0; python_version >= "3.7"
brotli==1.0.9; python_version >= "3.6" and python_full_version >= "3.8.0"
cchardet==2.1.7; python_version >= "3.6" and python_full_version >= "3.8.0"
certifi==2022.6.15; python_version >= "3.6"
cffi==1.15.1
charset-normalizer==2.1.0; python_version >= "3.6" and python_full_version >= "3.8.0"
discord.py @ git+https://github.com/Rapptz/discord.py.git@master ; python_full_version >= "3.8.0"
dnspython==2.2.1; python_version >= "3.7" and python_version < "4.0"
frozenlist==1.3.0; python_version >= "3.7" and python_full_version >= "3.8.0"
h11==0.9.0; python_version >= "3.6"
httpcore==0.11.1; python_version >= "3.6"
httptools==0.4.0; python_full_version >= "3.5.0" and python_version >= "3.6"
httpx==0.15.4; python_version >= "3.6"
idna==3.3; python_version >= "3.6" and python_full_version >= "3.8.0"
multidict==5.2.0; python_version >= "3.6" and python_full_version >= "3.8.0"
mutagen==1.45.1; python_version >= "3.6" and python_version < "4"
numpy==1.23.1; python_version >= "3.8"
psutil==5.9.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
pycares==4.2.1; python_version >= "3.6" and python_full_version >= "3.8.0"
pycparser==2.21; python_full_version >= "3.8.0" and python_version >= "3.6"
pycryptodomex==3.15.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
pydub==0.25.1
pymongo==4.2.0; python_version >= "3.7"
pynacl==1.5.0; python_full_version >= "3.8.0" and python_version >= "3.6"
python-engineio==4.3.3; python_version >= "3.6"
python-socketio==5.7.1; python_version >= "3.6"
rfc3986==1.5.0; python_version >= "3.6"
sanic==20.12.6; python_version >= "3.6"
sniffio==1.2.0; python_version >= "3.6"
typing-extensions==4.3.0; python_version >= "3.7"
ujson==5.4.0; sys_platform != "win32" and implementation_name == "cpython" and python_version >= "3.7"
uvloop==0.16.0; sys_platform != "win32" and implementation_name == "cpython" and python_version >= "3.7"
websockets==9.1; python_full_version >= "3.6.1" and python_version >= "3.6"
yarl==1.7.2; python_version >= "3.6" and python_full_version >= "3.8.0"
yt-dlp==2021.12.27; python_version >= "3.6"
